+++ CHANGES.md
@@ -56,6 +56,7 @@ Changed:
 - Deprecated `get_mime`, added `file.mime.libmagic` and `file.mime.cli`, made
   `file.mime` try `file.mime.libmagic` if present and `file.mime.cli` otherwise,
   changed eturned value when no mime was found to `null()`.
+- Return a nullable float in `request.duration`.
 - Removed `--list-plugins-json` and `--list-plugins-xml` options.
 - Added `--list-functions-json` option.
 - Removed built-in use of `strftime` conversions in output filenames, replaced